feat: Enforce integer validation for train_length and test_length in PUT request payload

+++ src/app/api/data-split/route.ts
@@ -6,8 +6,8 @@ import { z } from 'zod'
 
 export async function PUT(request: NextRequest) {
   const payload = z.object({
-    train_length: z.number(),
-    test_length: z.number(),
+    train_length: z.number().int(),
+    test_length: z.number().int(),
     reference: z.string()
   }).safeParse(await request.json())
